refactor: organize level loading login into its own file
